+++ fabric/changelog.md
@@ -1,2 +1,3 @@
 - Improved tooltip rendering performance
 - Fixed ping buffer size
+- Make permission manager more resilient against errors